Renamed module to kubos-hal-stm32f4. Added target dependencies for 405rg and 407vg variants.
