+++ module.json
@@ -1,16 +1,23 @@
 {
-  "name": "kubos-hal-stm32f407vg",
+  "name": "kubos-hal-stm32f4",
   "version": "0.0.1",
   "description": "KubOS - Satellite flight software platform",
   "keywords": [],
   "author": "kubos.co",
-  "homepage": "https://www.kubos.co",
+  "homepage": "https://github.com/openkosmosorg/kubos-hal-stm32f4",
   "license": "Apache-2.0",
   "extraIncludes": [
-    "kubos-hal-stm32f407vg"
+    "kubos-hal-stm32f4"
   ],
   "dependencies": {
-    "kubos-hal" : "openkosmosorg/kubos-hal#~0.0.1",
-    "stm32cubef4-stm32f407vg": "openkosmosorg/stm32cubef4-stm32f407vg#~0.0.1"
+    "kubos-hal" : "openkosmosorg/kubos-hal#~0.0.1"
+  },
+  "targetDependencies": {
+    "stm32f407vg": {
+      "stm32cubef4-stm32f407vg": "openkosmosorg/stm32cubef4-stm32f407vg#~0.0.1"
+    },
+    "stm32f405rg": {
+      "stm32cubef4-stm32f405rg": "openkosmosorg/stm32cubef4-stm32f405rg#master"
+    }
   }
 }
