+++ source/uart.c
@@ -22,7 +22,7 @@
    */
 #include "kubos-hal/gpio.h"
 #include "kubos-hal/uart.h"
-#include "kubos-hal-stm32f407vg/stm32f4_gpio.h"
+#include "kubos-hal-stm32f4/stm32f4_gpio.h"
 
 static inline USART_TypeDef *uart_dev(KUARTNum uart)
 {
